requirements.txt
# Telegram Bot Framework
aiogram==3.4.1

# YouTube downloader
yt-dlp==2023.12.30

# Environment variables
python-dotenv==1.0.0

# Database
aiosqlite==0.19.0

# HTTP client (dependency for aiogram)
aiohttp==3.9.1

# Audio processing (для yt-dlp)
ffmpeg-python==0.2.0

# Additional utilities
pathlib2==2.3.7.post1 